requirements.txt
google-cloud-texttospeech
pathlib
pydub
tqdm